[1-0005: Create REST to get menu of restaurant] fix bugs
