+++ epam-vote-system-controller/src/main/java/com/epam/vote/controller/RestaurantController.java
@@ -1,6 +1,8 @@
 package com.epam.vote.controller;
 
+import com.epam.vote.domain.dto.MenuDto;
 import com.epam.vote.domain.dto.RestaurantDto;
+import com.epam.vote.service.impl.MenuService;
 import com.epam.vote.service.impl.RestaurantService;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,12 +32,25 @@ public class RestaurantController {
     @Autowired
     private RestaurantService restaurantService;
 
+    @Autowired
+    private MenuService menuService;
+
     @GetMapping
     public ResponseEntity<List<RestaurantDto>> getAllRestaurants() {
         List<RestaurantDto> restaurantDtoList = restaurantService.getAllRestaurants().stream()
             .map(RestaurantDto::new).collect(Collectors.toList());
         return !restaurantDtoList.isEmpty()
             ? new ResponseEntity<>(restaurantDtoList, HttpStatus.OK)
+            : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping(value = "/{id}/menu")
+    public ResponseEntity<List<MenuDto>> getMenuOfRestaurant(
+        @PathVariable(name = "id") String id) {
+        List<MenuDto> menuDtoList = menuService.getMenuOfRestaurant(id).stream()
+            .map(MenuDto::new).collect(Collectors.toList());
+        return !menuDtoList.isEmpty()
+            ? new ResponseEntity<>(menuDtoList, HttpStatus.OK)
             : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
